+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG for PHP CS Fixer: custom fixers
 
 ## v3.32.0
+- Add FunctionParameterSeparationFixer
 - Update minimum PHP CS Fixer version to 3.85.0
 
 ## v3.31.0